+++ remotestate-ts/CHANGES.md
@@ -11,6 +11,8 @@
     action/query handlers as fallback clients.
   - Made `parsePath(path: string)` function public to support implementing 
     local stores using `createLocalRemoteStateClient()`.
+  - Added `Store.set(path, value)` and changed `useRemoteState()` writes to use
+    the store directly.
   - Exported `RemoteStateProviderProps`.
   - Updated docs with:
     - a basic counter RemoteState usage example